requirements.txt
selenium~=4.16.0
pytest~=7.4.4
webdriver-manager~=4.0.1
pytest-rerunfailures
pytest-pyreport
allure-pytest